remove GloballyAddressable class, move implementation to AGloballyAddressable
